requirements.txt
deepface==0.0.93
Django==5.1.2
djangorestframework==3.15.2